vim: complementary plugin to surround.vim
If you've ever tried using the "." command after a plugin map you were likely
disappointed to discover it only repeated the last native command inside that
app, rather than the map as a whole.  That disappointment ends today.
Repeat.vim remaps "." in a way that plugins can tap into it.