Remove ssh-agent handling.
Since SSH key handling depends quite on the system, I remove it from the core
startup. I almost exclusively work with Darwin environments that also comes
with a nice ssh-agent frontend that selectively loads the required keys. The
previous funtoo keychain solution loaded *all* keys in memory, which is not
necessary and should be avoided (principle of least privilege).