Rework mutt & offlineimap setup
Finally, after much agony, back to mutt for email handling. Offlineimap
got a lot smarter since I abonded it; the new native support for GMail
is a bliss. On the same token, this was a great chance to check some
basic Mutt workflow assumptions, so I reworked the key bindings to be
more memorable and efficient.